requirements.txt
Django == 5.0.6
djangorestframework == 3.15.2
psycopg2 == 2.9.9
psycopg2-binary == 2.9.9
django-environ == 0.11.2
clarifai_grpc == 10.5.4
django-cors-headers == 4.4.0
djangorestframework-simplejwt == 5.3.1
drf-spectacular == 0.27.1
drf-yasg == 1.21.7
setuptools == 69.5.1
packaging == 24.0